In operator add and sub, cast float to int64 before math with timestamp.

+++ r_exec/operator.cpp
@@ -292,7 +292,7 @@ namespace	r_exec{
 
 				if(lhs[0]!=Atom::PlusInfinity()){
 
-					index=context.setTimestampResult(Utils::GetTimestamp(&rhs[0])+lhs[0].asFloat());
+					index=context.setTimestampResult(Utils::GetTimestamp(&rhs[0])+(int64)lhs[0].asFloat());
 					return	true;
 				}
 			}
@@ -306,7 +306,7 @@ namespace	r_exec{
 
 				if(rhs[0]!=Atom::PlusInfinity()){
 
-					index=context.setTimestampResult(Utils::GetTimestamp(&lhs[0])+rhs[0].asFloat());
+					index=context.setTimestampResult(Utils::GetTimestamp(&lhs[0])+(int64)rhs[0].asFloat());
 					return	true;
 				}
 			}
@@ -352,7 +352,7 @@ namespace	r_exec{
 
 				if(rhs[0]!=Atom::PlusInfinity()){
 
-					index=context.setTimestampResult(Utils::GetTimestamp(&lhs[0])-rhs[0].asFloat());
+					index=context.setTimestampResult(Utils::GetTimestamp(&lhs[0])-(int64)rhs[0].asFloat());
 					return	true;
 				}
 			}
